@@ -2,124 +2,121 @@
   "name": "react-ts-playground",
   "version": "0.0.0",
   "description": "Playground for using react with typescript",
-  "author": "Christian Linne",
-  "license": "MIT",
+  "main": "index.js",
+  "repository": "git@github.com:DorianGrey/react-ts-playground.git",
+  "author": "Christian Linne <ShadowVampire@web.de>",
+  "license": "Unlicense",
   "private": true,
-  "dependencies": {
-    "immutable": "4.0.0-rc.9",
-    "lodash-es": "^4.17.10",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-intl": "^2.4.0",
-    "react-intl-redux": "^0.7.0",
-    "react-loadable": "^5.4.0",
-    "react-md": "^1.3.1",
-    "react-redux": "^5.0.7",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "redux": "^3.7.2",
-    "redux-observable": "^0.18.0",
-    "rxjs": "^5.5.6",
-    "webfontloader": "^1.6.28",
-    "whatwg-fetch": "2.0.4"
-  },
   "devDependencies": {
-    "@types/enzyme": "^3.1.10",
-    "@types/enzyme-adapter-react-16": "^1.0.2",
-    "@types/jest": "^22.2.3",
-    "@types/lodash-es": "^4.14.6",
-    "@types/node": "10.1.0",
-    "@types/react": "16.3.14",
-    "@types/react-dom": "^16.0.5",
-    "@types/react-hot-loader": "^3.0.6",
-    "@types/react-intl": "^2.3.8",
-    "@types/react-intl-redux": "^0.1.9",
-    "@types/react-loadable": "5.4.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@intervolga/optimize-cssnano-plugin": "^1.0.6",
+    "@types/enzyme": "^3.1.13",
+    "@types/enzyme-adapter-react-16": "^1.0.3",
+    "@types/jest": "^23.3.1",
+    "@types/lodash-es": "^4.17.1",
+    "@types/react": "^16.4.13",
+    "@types/react-dom": "^16.0.7",
+    "@types/react-intl": "^2.3.10",
+    "@types/react-intl-redux": "^0.1.10",
+    "@types/react-loadable": "^5.4.1",
     "@types/react-redux": "^5.0.18",
-    "@types/react-router": "^4.0.25",
-    "@types/react-router-dom": "^4.2.6",
+    "@types/react-router": "^4.0.30",
+    "@types/react-router-dom": "^4.3.0",
     "@types/webfontloader": "^1.6.29",
     "@types/webpack-env": "^1.13.6",
-    "app-root-path": "^2.0.1",
-    "autodll-webpack-plugin": "^0.3.9",
-    "autoprefixer": "^8.5.0",
+    "autoprefixer": "^9.1.3",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-loader": "^8.0.2",
+    "cache-loader": "^1.2.2",
     "case-sensitive-paths-webpack-plugin": "^2.1.2",
-    "chalk": "^2.4.1",
-    "cli-highlight": "2.0.0",
-    "compression": "^1.7.2",
-    "cross-env": "^5.1.5",
-    "css-loader": "^0.28.11",
-    "dotenv": "5.0.1",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.11",
-    "fork-ts-checker-webpack-plugin": "^0.3.0",
-    "fs-extra": "^6.0.1",
-    "glob": "^7.1.2",
-    "hard-source-webpack-plugin": "^0.5.18",
-    "html-webpack-plugin": "^2.30.1",
-    "http-proxy": "^1.17.0",
-    "husky": "^0.14.3",
-    "inline-chunk-manifest-html-webpack-plugin": "^2.0.0",
-    "jest": "^22.4.3",
-    "js-yaml": "^3.11.0",
-    "lint-staged": "^7.1.0",
+    "clean-webpack-plugin": "^0.1.19",
+    "copy-webpack-plugin": "^4.5.2",
+    "css-loader": "^1.0.0",
+    "detect-port": "^1.2.3",
+    "enzyme": "^3.5.0",
+    "enzyme-adapter-react-16": "^1.3.1",
+    "file-loader": "^2.0.0",
+    "fork-ts-checker-webpack-plugin": "^0.4.9",
+    "friendly-errors-webpack-plugin": "^1.7.0",
+    "fs-extra": "^7.0.0",
+    "hard-source-webpack-plugin": "^0.12.0",
+    "html-webpack-plugin": "^3.2.0",
+    "internal-ip": "^3.0.1",
+    "jest": "^23.5.0",
+    "jest-junit": "^5.1.0",
+    "jest-watch-typeahead": "^0.2.0",
+    "lint-staged": "^7.2.2",
     "lodash": "^4.17.10",
-    "log4js": "^2.6.1",
+    "mini-css-extract-plugin": "^0.4.2",
     "mkdirp": "^0.5.1",
-    "name-all-modules-plugin": "^1.0.1",
-    "node-sass": "^4.9.0",
-    "object-assign": "4.1.1",
-    "postcss-flexbugs-fixes": "^3.3.1",
-    "postcss-loader": "^2.1.5",
-    "prettier": "^1.12.1",
-    "promise": "^8.0.1",
-    "raf": "^3.4.0",
-    "react-dev-utils": "^5.0.1",
-    "react-error-overlay": "^4.0.0",
-    "react-hot-loader": "3.1.3",
+    "node-object-hash": "^1.4.1",
+    "node-sass": "^4.9.3",
+    "postcss-flexbugs-fixes": "^4.1.0",
+    "postcss-loader": "^3.0.0",
+    "prettier": "^1.14.2",
+    "readline": "^1.3.0",
+    "register-service-worker": "^1.5.2",
     "resolve-url-loader": "^2.3.0",
-    "rimraf": "^2.6.2",
-    "sass-loader": "^7.0.1",
-    "script-ext-html-webpack-plugin": "^1.8.8",
-    "source-map-loader": "^0.2.3",
-    "style-loader": "^0.21.0",
-    "stylelint": "^9.2.1",
+    "sass-loader": "^7.1.0",
+    "script-ext-html-webpack-plugin": "^2.0.1",
+    "signale": "^1.2.1",
+    "source-map-loader": "^0.2.4",
+    "style-loader": "^0.23.0",
+    "stylelint": "^9.5.0",
     "stylelint-config-standard": "^18.2.0",
-    "stylelint-scss": "^3.1.0",
-    "stylelint-webpack-plugin": "^0.10.4",
-    "sw-precache-webpack-plugin": "^0.11.5",
-    "ts-jest": "^22.4.6",
-    "ts-loader": "^3.5.0",
-    "tslib": "^1.9.1",
-    "tslint": "^5.10.0",
-    "tslint-config-prettier": "^1.12.0",
+    "stylelint-scss": "^3.3.0",
+    "terser-webpack-plugin": "^1.0.2",
+    "ts-jest": "^23.1.4",
+    "ts-loader": "^5.0.0",
+    "tslib": "^1.9.3",
+    "tslint": "^5.11.0",
+    "tslint-config-prettier": "^1.15.0",
     "tslint-react": "^3.6.0",
-    "typescript": "2.8.3",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.11.0",
-    "webpack-bundle-analyzer": "^2.11.1",
-    "webpack-chunk-hash": "^0.5.0",
-    "webpack-common-shake": "^1.5.3",
-    "webpack-dev-server": "^2.11.1",
-    "webpack-manifest-plugin": "^1.3.2",
-    "webpack-merge": "^4.1.2",
-    "yargs": "^11.0.0"
+    "tslint-sonarts": "^1.7.0",
+    "typescript": "^3.0.3",
+    "url-loader": "^1.1.1",
+    "webpack": "^4.17.1",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-chain": "^4.10.0",
+    "webpack-chunk-hash": "^0.6.0",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.8",
+    "workbox-webpack-plugin": "^3.4.1",
+    "yorkie": "^2.0.0"
+  },
+  "dependencies": {
+    "lodash-es": "^4.17.10",
+    "react": "^16.5.0",
+    "react-dom": "^16.5.0",
+    "react-intl": "^2.4.0",
+    "react-intl-redux": "^2.0.2",
+    "react-loadable": "^5.5.0",
+    "react-md": "^1.5.1",
+    "react-redux": "^5.0.7",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
+    "redux": "^4.0.0",
+    "redux-observable": "^1.0.0",
+    "rxjs": "^6.3.1",
+    "webfontloader": "^1.6.28"
   },
   "scripts": {
-    "prestart": "npm run translations && npm run renderLoadingAnimation",
-    "start": "node scripts/start.js",
-    "prebuild": "npm run translations && npm run renderLoadingAnimation",
-    "build": "node scripts/build.js",
+    "start": "webpack-dev-server --config config/webpack/resolvedConfig.dev.js",
+    "build": "webpack --config config/webpack/resolvedConfig.prod.js",
     "serve": "node scripts/serve.js",
-    "test:ci": "node scripts/test.js --env=jsdom --no-cache --verbose --coverage",
-    "test": "node scripts/test.js --env=jsdom --verbose",
-    "renderLoadingAnimation": "node scripts/util/renderLoading.js",
+    "test:ci": "node scripts/test.js --no-cache --coverage",
+    "test": "node scripts/test.js",
+    "inspect": "node scripts/inspect.js",
     "format": "prettier --write \"{src,scripts,config}/**/*.{js,ts,tsx}\"",
-    "precommit": "lint-staged",
-    "prepush": "npm run test:ci",
-    "translations": "node scripts/translations/compile-translations.js \"src/**/*.i18n.yml\" src/generated/translations -s"
+    "lint": "tslint -c tslint.json -p tsconfig.json --fix && stylelint \"src/**/*.scss\""
+  },
+  "gitHooks": {
+    "pre-commit": "yarn lint && lint-staged",
+    "pre-push": "yarn test:ci"
   },
   "lint-staged": {
     "{src,scripts,config}/**/*.{js,ts,tsx}": [
