@@ -1,4 +1,4 @@
-import * as Enzyme from "enzyme";
-import * as Adapter from "enzyme-adapter-react-16";
+import Enzyme from "enzyme";
+import Adapter from "enzyme-adapter-react-16";
 
 Enzyme.configure({ adapter: new Adapter() });
