@@ -3,16 +3,13 @@
 const express = require("express");
 const compression = require("compression");
 const path = require("path");
-const logger = require("log4js").getLogger("server");
-logger.level = "debug";
-
-const httpProxy = require("http-proxy");
+const logger = require("signale");
 
 // Enter your proxy rules here.
 const serverPort = 4000;
 const app = express();
 const router = express.Router();
-const proxy = httpProxy.createProxyServer();
+
 // Configure proxy as you'd like here...
 
 const serveDirs = process.argv.slice(2);
