@@ -3,6 +3,8 @@
 process.env.NODE_ENV = "test";
 process.env.PUBLIC_URL = "";
 
+const compileTranslations = require("./translations").compile;
+
 // Makes the script crash on unhandled rejections instead of silently
 // ignoring them. In the future, promise rejections that are not handled will
 // terminate the Node.js process with a non-zero exit code.
@@ -10,15 +12,22 @@ process.on("unhandledRejection", err => {
   throw err;
 });
 
-// Ensure environment variables are read.
-require("../config/env");
+async function runTest() {
+  await compileTranslations("src/**/*.i18n.yml", "src/generated/translations", {
+    splitPerLang: true
+  });
 
-const jest = require("jest");
-const argv = process.argv.slice(2);
+  const jest = require("jest");
+  const argv = process.argv.slice(2);
 
-// Watch unless on CI or in coverage mode
-if (!process.env.CI && argv.indexOf("--coverage") < 0) {
-  argv.push("--watch");
+  // Watch unless on CI or in coverage mode
+  if (!process.env.CI && argv.indexOf("--coverage") < 0) {
+    argv.push("--watch");
+  }
+
+  jest.run(argv);
 }
 
-jest.run(argv);
+runTest().catch(err => {
+  console.error(err);
+});
