@@ -1,5 +1,5 @@
 import { mount, shallow } from "enzyme";
-import * as React from "react";
+import React from "react";
 
 import NotFound from "./404";
 
