@@ -1,4 +1,4 @@
-import * as en from "react-intl/locale-data/en";
+import en from "react-intl/locale-data/en";
 import translationsEn from "../../../generated/translations.en";
 import { LanguagePack } from "./languagePack";
 
